none

+++ makefile
@@ -1,6 +1,6 @@
 # vim: noet sw=8 sts=8
-DOTFILES= ${shell for i in .??* *; do case $$i in \
-	  .git|.gitignore|makefile);; \
+LINKS	= ${shell for i in .* *; do case $$i in \
+	  .|..|.git|.gitignore|makefile);; \
 	  *) echo $$i;; esac; done}
 DOTDIR	= $(shell pwd | sed -e "s:^$$HOME/::")
 PRIVATE	= $(HOME)/private/dot
@@ -15,8 +15,9 @@ commit:
 	git push
 
 relink:
-	@for f in $(DOTFILES); do \
+	for f in $(LINKS); do \
 	    if [ -L $(HOME)/$$f ]; then rm $(HOME)/$$f || return 1; fi; \
+	    if [ -e $(HOME)/$$f ]; then return 1; fi; \
 	    ln -s $(DOTDIR)/$$f $(HOME) || return 1; \
 	    find $(HOME)/$$f -printf '%p -> %l\n'; \
 	done
